* CET-283: Importing Text With Carriage Returns Confuses Record Count (#32610)

* CET-271: Published Pre-Set imports not appearing for Users (#32742)

* PAT-2745 (#32741)

* CET-211: Assistance with HotFix for Bug 76746 (#32666)

* CET-279: Double Primary Teams When One Team Is Global (ID 1) And The Other's ID starts with 1 (#32613)

* CET-329-Latin alphabet display issue in help text (#32901)

* CET-320: Editing Contract inline in Accounts' Contracts subpanel unlinks Contract from Opportunity (#32855)

* CET-310: Reports - Autocomplete is not triggered from a Filter (#32858)

* CET-313: After rename, old name of a module is appearing in reports module (#32842)

* CET-305:Visibility Editor may create URL's beyond the server's length limit (#32775)

* Revert "[BR-3525] (SI 74590) "Cross Site Request Forgery (XSRF) Attack Detected" Message Appears in Studio / Module Builder for some actions"

This reverts commit 0d8faa8bdf36526d560f5cfe0f0db546e746a880.

* CET-258: Visibility Editor may create URL's beyond the server's length
limit

* CET-273: When account and contact has same ID and the contact is in the target list, it will generate the wrong export target csv file (#32863)

* CET-269: Subpanels Broken after Upgrade (#32857)

* CET-285 to handle sub module module_name (#32603)

* CET-295: Roles persist for user after removal (#32745)

deleted

properly

* CET-227: namespaces - upgrading in v7.7 causes logic hook files to be (#32786)

deleted

* CET-316: Help Text Does not display properly characters with umlauts (#32956)

* CET-357: Fix Code Standard Errors

+++ sugarcrm/modules/Contracts/Contract.php
@@ -143,8 +143,11 @@ class Contract extends SugarBean
         // contracts_opportunities is many to many but for some reason the UI implements it as if it's one to many
         // that is, for a contract, only one opp can be linked to it
         // workaround here to delete the current entry so when it's inserted later we have only the new record
-        $query = 'delete from ' . $this->rel_opportunity_table . " where contract_id = '" . $this->id . "'";
-        $this->db->query($query);
+        if (!empty($this->rel_fields_before_value['opportunity_id']) &&
+            $this->rel_fields_before_value['opportunity_id'] != $this->opportunity_id) {
+            $query = 'delete from ' . $this->rel_opportunity_table . " where contract_id = '" . $this->id . "'";
+            $this->db->query($query);
+        }
 
         $this->setCalculatedValues(false);
         $return_id = parent::save($check_notify);
